packages/backend/pyproject.toml
[project]
name = "newsletter-backend"
version = "0.1.0"
description = "Newsletter backend — fetches content and sends to email-service via gRPC"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "feedparser>=6.0.0",
    "yfinance>=0.2.0",
    "astral>=3.2",
    "deep-translator>=1.11.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "mypy-protobuf>=3.6.0",
    "grpc-stubs>=1.53.0",
]

[project.scripts]
newsletter = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "generated"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "mypy-protobuf>=3.6.0",
    "grpc-stubs>=1.53.0",
    "types-pyyaml>=6.0.12.20250915",
]
